[State Optimization] Refactor MsaKeysOf to store a reference counter (#444)
# Goal
The goal of this PR is to save on-chain data by refactoring the
MsaKeysOf storage map to no longer store the mapping of MSA ids to
account keys. We have found that we do not need to store the mapping of
MSA Ids to associated keys. Instead, we can just count references.

Closes #384

# Discussion
RPC `get_msa_keys` and references to MSA helper function
`fetch_msa_keys` have been temporarily removed until #418 at which time
they should be reimplemented to retrieve keys associated to an MSA with
`MessageSourceIdOf`.

# Checklist
- [x] Updated the js/api-augment code if a custom RPC added/changed
- [x] Design doc(s) updated
- [x] Tests update