Chores/common runtime config (#479)
# Goal
The goal of this PR is to refactor common settings between frequency and
frequency-rococo chains into their own module and use where needed, so
as to prevent copy-paste errors and easily keep the two chains aligned
in configuration where they need to be.

Closes #375 

# Discussion
This should not change any behavior. I also made a couple of numerically
named constants for easier reading. Some constants/types could not
easily be unified and moved into the new common files because they had
too many dependencies on the runtime-specific config.

# Checklist
- [ ] Chain spec updated
- [x] N/A Updated the js/api-augment code if a custom RPC added/changed
- [x] N/A Design doc(s) updated
- [x] N/A Tests added
- [x] N/A Benchmarks added