Chores/review unwrap #561 (#584)
# Goal
The goal of this PR is to ensure that there are no hidden possible
panics in production code.

Closes #561 

# Discussion

- [x] common/helpers/src/avro.rs converted to use `match`
- [x] pallets/messages/src/lib.rs - Two check first for `is_some()` (and
converting them to matches makes the code less readable) so they're
fine. The other is now `unwrap_or_default` with an empty check.
- [x] pallets/messages/src/types.rs (completed by @saraswatpuneet )
- [x] pallets/messages/src/rpc/src/lib.rs - converted to
`unwrap_or_default()`
- [x] pallets/msa/src/lib.rs - this one is only for benchmarks
- [x] pallets/msa/src/rpc/src/lib.rs - replace with `unwrap_or(false)`
- [x] runtime/frequency/src/lib.rs - - this one is only for
`try-runtime` and not production
- [x] runtime/frequency-rococo/src/lib.rs - this one is only for
`try-runtime` and not production

## Additionally
Code coverage was reduced in `avro.rs` so I had to write/rewrite some
tests. In the process I reorganized the tests so that valid schemas and
invalid schemas are completely separate, and tests for schema errors are
separate from happy path tests.

# Checklist
- [x] N/A Chain spec updated
- [x] N/A Updated the js/api-augment code if a custom RPC added/changed
- [x] N/A Design doc(s) updated
- [x] Tests added
- [X] N/A Benchmarks added
- [X] N/A Weights updated