upload encoded metadata to artifacts (#725)
# Goal
The goal of this PR is to add a new CI job to the "Merge PR" workflow
which would extract the encoded metadata and upload it to the GitHub
artifacts. Each upload will override the previous one, so we always have
the latest `main`. Later, we'll download this metadata and use it as a
reference when comparing the metadata from the freshly built binary in
the Verify PR workflow.

Closes #715