fix(cid-as-string): Return CID as string instead of byte array #787 (#1124)
# Goal
The goal of this PR is to return CID as a string instead of `Vec<u8>`
bytes. Based on:

> Peter Frank
[6:36
PM](https://unfinished-team.slack.com/archives/C01L4J5SQA1/p1675467404682619)
@here think I found a bug, or really an incomplete ticket. 
[@Aziz Abada](https://unfinished-team.slack.com/team/U02J2JHFZJ4)
is integrating against the messages.getMessagesBySchemaId call and for
the cid itself we are being returned a byte array (really a JSON array
of very small numbers that are less than u8 :)). It appears this was
noticed in this ticket
https://github.com/LibertyDSNP/frequency/issues/787 and the comment on
this code
https://github.com/LibertyDSNP/frequency/blob/main/common/primitives/src/messages.rs#L38-L39
talks about the base32 encoding in the comment but it doesn’t actually
change the data type from Option<Vec<u8>> to Option<string>. Just wanted
to call this out, we will be adding a method to convert this on our side
for now receiving a naked byte array (but especially one that is really
a string) isn’t terribly useful but not sure how we go about
fixing/changing this.

Closes #787 

# Checklist
- [ ] Chain spec updated
- [ ] Custom RPC OR Runtime API added/changed? Updated js/api-augment.
- [ ] Design doc(s) updated
- [ ] Tests added
- [ ] Benchmarks added
- [ ] Weights updated