offchain: reverse mapping from msa to keys (#1810)
# Goal
This PR adds the feature of being able to get all the keys for a certain
MSA and generally implements the infra structure to be able to do
indexing of desired data on offchain DB.

Closes #418 

# Diagram


![reverse_mapping](https://github.com/LibertyDSNP/frequency/assets/9152501/137eedb8-ed02-44ae-918b-458f45be57a6)

# Discussions
- Offchain worker gets the finalized block with an http call to it's own
rpc using `http://localhost:9944`. This means that the node should
expose RPCs (at least locally). This is based on
https://github.com/paritytech/substrate/issues/6742. There might be a
way to be able to pass finalized block into offchain worker without RPC
call but was not able to identify it so far.
- Offchain worker runs by default the following condition so we need to
explicitly disallow them in cases that we don't want them to run by
default
`- when-authority: Only enable the offchain worker when running as a
validator (or collator, if this is a parachain node)`
- There is an inherent assumption that accessing offchain indexed data
after finalization should prevent us from dealing with fork situations
since the canonical block should be the last block getting imported for
all finalized ones. More info in
https://substrate.stackexchange.com/questions/10757/offchain-indexing-storage-overwrites-during-forks
- We might want to disallow 'offchain_localStorageSet' RPC to ensure the
indexed data would not be altered easily.


# Checklist
- [X] Chain spec updated
- [x] Update running scripts
- [x] Custom RPC OR Runtime API added/changed? Updated js/api-augment.
- [X] Tests added